+++ tester/action/errors.go
@@ -60,7 +60,8 @@ var (
 	ErrActionAssertLte           = newErrAction("AssertLte")
 	ErrActionAssertAll           = newErrAction("AssertAll")
 	ErrActionAssertAny           = newErrAction("AssertAny")
-	ErrActionAssertRegex         = newErrAction("AssertRegex")
+	ErrActionAssertRegexMatch    = newErrAction("ErrActionAssertRegexMatch")
+	ErrActionAssertRegexSearch   = newErrAction("ErrActionAssertRegexSearch")
 	ErrActionUnmarshal           = newErrAction("Unmarshal")
 	ErrActionUnmarshalObjectKeys = newErrAction("UnmarshalObjectKeys")
 )
