+++ tester/action/zero.go
@@ -1,0 +1,50 @@
+/*
+ * MIT License
+ *
+ * Copyright (c) 2025 Peter Vrba
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+package action
+
+import "encoding/json"
+
+// zeroValue is a type that represents a zero value in JSON.
+// It is used to compare against the zero values of raw json messages.
+type zeroValue json.RawMessage
+
+var (
+	zeroValueNumeric zeroValue = zeroValue("0")
+	zeroValueString            = zeroValue(`""`)
+	zeroValueBool    zeroValue = zeroValue("false")
+	zeroValueArray   zeroValue = zeroValue("[]")
+	zeroValueStruct  zeroValue = zeroValue("{}")
+	zeroValueNil     zeroValue = zeroValue("null")
+
+	// zeroValues is a list of all zero values
+	zeroValues = []zeroValue{
+		zeroValueNumeric,
+		zeroValueString,
+		zeroValueBool,
+		zeroValueArray,
+		zeroValueStruct,
+		zeroValueNil,
+	}
+)
